Delete Project implemented

+++ business/facade/ProjectFacade.java
@@ -2,14 +2,17 @@ package business.facade;
 
 import DAO.ProjectDAO;
 import DAO.factory.AbstractFactoryDAO;
-import business.system.*;
+import business.system.Collaborator;
+import business.system.Project;
+import business.system.Sprint;
+import business.system.User;
 
 import java.sql.Date;
 import java.sql.SQLException;
 import java.util.List;
 
 public class ProjectFacade {
-    
+
     private GlobalFacade globalFacade;
 
     public List<Project> getProjectListFromUser(User user) throws SQLException {
@@ -21,7 +24,6 @@ public class ProjectFacade {
         ProjectDAO projectDAO = AbstractFactoryDAO.getInstance().createProjectDAO();
         return projectDAO.getProjectByID(idProject);
     }
-
 
 
     public List<Collaborator> getProjectTeam(Project project) throws SQLException {
@@ -50,15 +52,14 @@ public class ProjectFacade {
     }
 
 
-
     public Project createProject(String name, String summary, String type, Date deadline, User user) throws SQLException {
         ProjectDAO projectDAO = AbstractFactoryDAO.getInstance().createProjectDAO();
         Project project = projectDAO.createProject(name, summary, type, deadline);
-        if(project == null){
+        if (project == null) {
             return null;
         }
         boolean success = globalFacade.initiateProductTicketBacklog(project.getId());
-        if(!success){
+        if (!success) {
             return null;
         }
         String role = "Scrum Master";
@@ -84,9 +85,33 @@ public class ProjectFacade {
     }
 
     public boolean deleteProject(Project project) throws SQLException {
+        // Remove Sprint
+        Sprint[] sprints = globalFacade.getSprintsByProject(project);
+        for (Sprint s : sprints) {
+            boolean success = globalFacade.deleteSprint(s);
+            if (!success) {
+                System.err.println("Pb delete Sprint");
+            }
+        }
+
+        // Remove Backlogs
+        boolean success = globalFacade.deleteBacklogs(project);
+        if (!success) {
+            System.err.println("Err Rm Backlogs");
+            return false;
+        }
+
+        // Remove Collaborators
+        List<Collaborator> projectTeam = getProjectTeam(project);
+        for (Collaborator c : projectTeam) {
+            removeCollaborator(c);
+        }
+
+        // delete Project
         ProjectDAO projectDAO = AbstractFactoryDAO.getInstance().createProjectDAO();
         return projectDAO.deleteProject(project.getId());
     }
+
